Change signature of sample to double, incorporate sample frequency into the RateControl

+++ common/src/main/java/com/salesforce/ouroboros/util/rate/Controller.java
@@ -57,7 +57,7 @@ public interface Controller {
      * @param responseTime
      *            - the new sample
      */
-    public void sample(int responseTime);
+    public void sample(double responseTime);
 
     /**
      * @return the median of the sampled response time over the history of the
